+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Added
 * Support for the new query parameters to find incoming links [to a specific entry](https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/search-parameters/links-to-entry) or [to a specific asset](https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/search-parameters/links-to-asset) in a space.
 
+### Fixed
+* Fixes coercion issue when `DateTime` values are `nil`. [#148](https://github.com/contentful/contentful.rb/issues/148)
+
 ## 2.2.2
 ### Fixed
 * Fixed 404 error when `details` object contains a top level `sys` key.